// projects/pos/src/app/shared/enums/columns/supplier-columns.enum.ts
export enum SupplierColumnLabel {
  ID = 'Id',
  NAME = 'Name',
  EMAIL = 'Email',
  PHONE = 'Phone',
  BALANCE = 'Balance',
  COMMENT = 'Comment',
  STATUS = 'Status',
}

export enum SupplierColumn {
  ID = 'id',
  NAME = 'name',
  EMAIL = 'email',
  PHONE = 'phone',
  BALANCE = 'balance',
  COMMENT = 'comment',
  STATUS = 'status',
}